Move filter by group test
